+++ src/forges.c
@@ -181,7 +181,7 @@ github_forge_descriptor =
 	                           | GCLI_MILESTONE_QUIRKS_DUEDATE
 	                           | GCLI_MILESTONE_QUIRKS_PULLS,
 	.pull_summary_quirks       = GCLI_PRS_QUIRK_COVERAGE
-	                           | GCLI_PRS_QUIRK_AUTOMERGE,
+	                           | GCLI_PRS_QUIRK_AUTOMERGE, /* automerge field seems broken */
 };
 
 static struct gcli_forge_descriptor const
@@ -383,6 +383,8 @@ gitea_forge_descriptor =
 	                           | GCLI_PRS_QUIRK_DRAFT
 	                           | GCLI_PRS_QUIRK_CHANGES
 	                           | GCLI_PRS_QUIRK_COVERAGE,
+	.pull_quirks               = GCLI_PULL_QUIRK_AUTOMERGE, /* uses Github Backend with GraphQL
+	                                                         * which doesn't work with gitea */
 };
 
 static struct gcli_forge_descriptor const