+++ src/pulls.c
@@ -151,8 +151,8 @@ int
 gcli_pull_submit(struct gcli_ctx *ctx, struct gcli_submit_pull_options *opts)
 {
 	if (opts->automerge) {
-		int const q = gcli_forge(ctx)->pull_summary_quirks;
-		if (q & GCLI_PRS_QUIRK_AUTOMERGE)
+		int const q = gcli_forge(ctx)->pull_quirks;
+		if (q & GCLI_PULL_QUIRK_AUTOMERGE)
 			return gcli_error(ctx, "forge does not support auto-merge");
 	}